Move cache classes to llama_cache submodule.
